requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
apscheduler
python-dotenv